rename package service to server
