+++ src/main/scala/nl/knaw/dans/easy/pid/server/ServerWiring.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid.service
+package nl.knaw.dans.easy.pid.server
 
 import nl.knaw.dans.easy.pid.ConfigurationComponent
 import nl.knaw.dans.easy.pid.generator.{ DOIGeneratorWiring, URNGeneratorWiring }
