+++ src/main/scala/nl/knaw/dans/easy/pid/server/ServletMounterComponent.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid.service
+package nl.knaw.dans.easy.pid.server
 
 import javax.servlet.ServletContext
