+++ src/test/scala/nl/knaw/dans/easy/pid/server/PidServerSpec.scala
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid.service
+package nl.knaw.dans.easy.pid.server
 
 import java.net.ConnectException
 
 import nl.knaw.dans.easy.pid.generator._
-import nl.knaw.dans.easy.pid.{ DatabaseAccessComponent, ConfigurationSupportFixture, SeedDatabaseFixture, ServerTestSupportFixture }
+import nl.knaw.dans.easy.pid.{ ConfigurationSupportFixture, DatabaseAccessComponent, SeedDatabaseFixture, ServerTestSupportFixture }
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.OneInstancePerTest
