+++ src/test/scala/nl/knaw/dans/easy/pid/server/PidServletSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid.service
+package nl.knaw.dans.easy.pid.server
 
 import java.sql.Connection
