+++ src/test/scala/nl/knaw/dans/easy/pid/server/ServerWiringSpec.scala
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid.service
+package nl.knaw.dans.easy.pid.server
 
-import nl.knaw.dans.easy.pid.generator._
 import nl.knaw.dans.easy.pid._
+import nl.knaw.dans.easy.pid.generator._
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
 import org.scalatest.OneInstancePerTest
